New SPA (fix connect button double click issue)
