+++ VpnHood.Client.App.Android.GooglePlay/Ads/InterstitialAdLoadCallback.cs
@@ -6,10 +6,9 @@ using Android.Runtime;
 namespace VpnHood.Client.App.Droid.GooglePlay.Ads;
 
 // fix google ad library problem
-public class InterstitialAdLoadCallback2(Activity activity) : InterstitialAdLoadCallback
+public class InterstitialAdLoadCallback(Activity activity) : global::Android.Gms.Ads.Interstitial.InterstitialAdLoadCallback
 {
     private static Delegate? _cbOnAdLoaded;
-    private InterstitialAd? _mInterstitialAd;
 
     // ReSharper disable once UnusedMember.Local
     private static Delegate GetOnAdLoadedHandler()
@@ -19,23 +18,15 @@ public class InterstitialAdLoadCallback2(Activity activity) : InterstitialAdLoad
 
     private static void OnAdLoadedNative(IntPtr env, IntPtr nativeThis, IntPtr nativeP0)
     {
-        var interstitialAdLoadCallback = GetObject<InterstitialAdLoadCallback2>(env, nativeThis, JniHandleOwnership.DoNotTransfer);
-        var interstitialAd = GetObject<InterstitialAd>(nativeP0, JniHandleOwnership.DoNotTransfer);
+        var interstitialAdLoadCallback = GetObject<InterstitialAdLoadCallback>(env, nativeThis, JniHandleOwnership.DoNotTransfer);
+        var interstitialAd = GetObject<global::Android.Gms.Ads.Interstitial.InterstitialAd>(nativeP0, JniHandleOwnership.DoNotTransfer);
         if (interstitialAd != null)
             interstitialAdLoadCallback?.OnAdLoaded(interstitialAd);
     }
 
     // ReSharper disable once StringLiteralTypo
     [Register("onAdLoaded", "(Lcom/google/android/gms/ads/interstitial/InterstitialAd;)V", "GetOnAdLoadedHandler")]
-    public virtual void OnAdLoaded(InterstitialAd interstitialAd)
+    public virtual void OnAdLoaded(global::Android.Gms.Ads.Interstitial.InterstitialAd interstitialAd)
     {
-        _mInterstitialAd = interstitialAd;
-        base.OnAdLoaded(interstitialAd);
-    }
-
-    public override void OnAdFailedToLoad(LoadAdError loadAdError)
-    {
-        _mInterstitialAd = null;
-        base.OnAdFailedToLoad(loadAdError);
     }
 }
